Add a snippet to redirect after login
The snippet is actually still necessary on the home page. CMS entry
page is fine without it. If is needed when a new user is invited. When
she clicks the link in invitation email, it redirects her to site’s
home page where invite token is verified and her new password is
set. After that, she’s logged in. This is where the snippet is
important, because it redirects her to the CMS.