Add project filtering implementation
It’s similar to user answer filtering, but not the same. The categories filter is exactly the same. But projects have date ranges (or no dates at all) and date filter can be either a year or a month (with year).

Project sorting has been moved to include which creates sorted_projects variable. The complexity of sorting increased and maintaing this algorithm at two places was already cumbersome.